Refactor sln into separate static libraries
