+++ Sources/ElgatoSD/EPLJSONUtils.h
@@ -17,7 +17,7 @@
 // Includes
 //------------------------------------------------------------------------------
 
-#include "../Vendor/json/src/json.hpp"
+#include "Vendor/json/src/json.hpp"
 using json = nlohmann::json;
 
 class EPLJSONUtils
