+++ Sources/ElgatoSD/ESDLocalizer.h
