+++ Sources/ElgatoSD/ESDSDKDefines.h
