+++ Sources/ElgatoSD/ESDUtilities.h
