+++ Sources/ElgatoSD/pch.h
@@ -44,19 +44,4 @@ void __cdecl dbgprintf(const char *format, ...);
 #endif
 
 
-//-------------------------------------------------------------------
-// json
-//-------------------------------------------------------------------
-
-#include "../Vendor/json/src/json.hpp"
-using json = nlohmann::json;
-
-
-//-------------------------------------------------------------------
-// websocketpp
-//-------------------------------------------------------------------
-
-#define ASIO_STANDALONE
-
-
 #endif //PCH_H
