+++ Sources/StreamdeckContext/SendActions/MomentaryContext.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "../StreamdeckContext.h"
+#include "StreamdeckContext/StreamdeckContext.h"
 
 #include <optional>
 #include <string>
