+++ Sources/StreamdeckContext/SendActions/SwitchContext.cpp
@@ -1,10 +1,8 @@
 // Copyright 2021 Charles Tytler
 
-#include "pch.h"
-
 #include "SwitchContext.h"
 
-#include "../../Common/EPLJSONUtils.h"
+#include "ElgatoSD/EPLJSONUtils.h"
 
 void SwitchContext::handleButtonEvent(DcsInterface &dcs_interface, const KeyEvent event, const json &inPayload)
 {
