+++ Sources/StreamdeckContext/StreamdeckContext.cpp
@@ -1,10 +1,8 @@
 // Copyright 2020 Charles Tytler
 
-#include "pch.h"
-
 #include "StreamdeckContext.h"
 
-#include "../Common/EPLJSONUtils.h"
+#include "ElgatoSD/EPLJSONUtils.h"
 
 StreamdeckContext::StreamdeckContext(const std::string &context) : context_{context} {};
