+++ Sources/Utilities/UdpSocket.cpp
@@ -2,8 +2,6 @@
 
 #pragma comment(lib, "Ws2_32.lib")
 
-#include "pch.h"
-
 #include "UdpSocket.h"
 
 #include <WS2tcpip.h>
