+++ Sources/Windows/com.ctytler.dcs.sdPlugin.sln
@@ -9,6 +9,32 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Test", "..\Test\Test.vcxpro
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Lua", "..\Vendor\lua-5.1.5\Lua.vcxproj", "{59EEC206-84E0-4842-B89E-2025FD78E589}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "StreamdeckContext", "..\StreamdeckContext\StreamdeckContext.vcxproj", "{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}"
+	ProjectSection(ProjectDependencies) = postProject
+		{F60CA831-72E9-4992-BCFE-12B3387F646B} = {F60CA831-72E9-4992-BCFE-12B3387F646B}
+		{2782D849-65F7-4B89-A856-964FD918269A} = {2782D849-65F7-4B89-A856-964FD918269A}
+		{329F64AC-6346-46D2-A671-9B5DD626B485} = {329F64AC-6346-46D2-A671-9B5DD626B485}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Utilities", "..\Utilities\Utilities.vcxproj", "{329F64AC-6346-46D2-A671-9B5DD626B485}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DcsInterface", "..\DcsInterface\DcsInterface.vcxproj", "{2782D849-65F7-4B89-A856-964FD918269A}"
+	ProjectSection(ProjectDependencies) = postProject
+		{59EEC206-84E0-4842-B89E-2025FD78E589} = {59EEC206-84E0-4842-B89E-2025FD78E589}
+		{329F64AC-6346-46D2-A671-9B5DD626B485} = {329F64AC-6346-46D2-A671-9B5DD626B485}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ElgatoSD", "..\ElgatoSD\ElgatoSD.vcxproj", "{F60CA831-72E9-4992-BCFE-12B3387F646B}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "StreamdeckInterface", "..\StreamdeckInterface\StreamdeckInterface.vcxproj", "{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}"
+	ProjectSection(ProjectDependencies) = postProject
+		{59EEC206-84E0-4842-B89E-2025FD78E589} = {59EEC206-84E0-4842-B89E-2025FD78E589}
+		{F60CA831-72E9-4992-BCFE-12B3387F646B} = {F60CA831-72E9-4992-BCFE-12B3387F646B}
+		{2782D849-65F7-4B89-A856-964FD918269A} = {2782D849-65F7-4B89-A856-964FD918269A}
+		{329F64AC-6346-46D2-A671-9B5DD626B485} = {329F64AC-6346-46D2-A671-9B5DD626B485}
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E} = {36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -41,6 +67,46 @@ Global
 		{59EEC206-84E0-4842-B89E-2025FD78E589}.Release|x64.Build.0 = Release|x64
 		{59EEC206-84E0-4842-B89E-2025FD78E589}.Release|x86.ActiveCfg = Release|Win32
 		{59EEC206-84E0-4842-B89E-2025FD78E589}.Release|x86.Build.0 = Release|Win32
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Debug|x64.ActiveCfg = Debug|x64
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Debug|x64.Build.0 = Debug|x64
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Debug|x86.ActiveCfg = Debug|Win32
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Debug|x86.Build.0 = Debug|Win32
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Release|x64.ActiveCfg = Release|x64
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Release|x64.Build.0 = Release|x64
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Release|x86.ActiveCfg = Release|Win32
+		{36A2B3D3-C3FA-424C-9BB6-37D19EBE5A3E}.Release|x86.Build.0 = Release|Win32
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Debug|x64.ActiveCfg = Debug|x64
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Debug|x64.Build.0 = Debug|x64
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Debug|x86.ActiveCfg = Debug|Win32
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Debug|x86.Build.0 = Debug|Win32
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Release|x64.ActiveCfg = Release|x64
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Release|x64.Build.0 = Release|x64
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Release|x86.ActiveCfg = Release|Win32
+		{329F64AC-6346-46D2-A671-9B5DD626B485}.Release|x86.Build.0 = Release|Win32
+		{2782D849-65F7-4B89-A856-964FD918269A}.Debug|x64.ActiveCfg = Debug|x64
+		{2782D849-65F7-4B89-A856-964FD918269A}.Debug|x64.Build.0 = Debug|x64
+		{2782D849-65F7-4B89-A856-964FD918269A}.Debug|x86.ActiveCfg = Debug|Win32
+		{2782D849-65F7-4B89-A856-964FD918269A}.Debug|x86.Build.0 = Debug|Win32
+		{2782D849-65F7-4B89-A856-964FD918269A}.Release|x64.ActiveCfg = Release|x64
+		{2782D849-65F7-4B89-A856-964FD918269A}.Release|x64.Build.0 = Release|x64
+		{2782D849-65F7-4B89-A856-964FD918269A}.Release|x86.ActiveCfg = Release|Win32
+		{2782D849-65F7-4B89-A856-964FD918269A}.Release|x86.Build.0 = Release|Win32
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Debug|x64.ActiveCfg = Debug|x64
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Debug|x64.Build.0 = Debug|x64
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Debug|x86.ActiveCfg = Debug|Win32
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Debug|x86.Build.0 = Debug|Win32
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Release|x64.ActiveCfg = Release|x64
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Release|x64.Build.0 = Release|x64
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Release|x86.ActiveCfg = Release|Win32
+		{F60CA831-72E9-4992-BCFE-12B3387F646B}.Release|x86.Build.0 = Release|Win32
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Debug|x64.ActiveCfg = Debug|x64
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Debug|x64.Build.0 = Debug|x64
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Debug|x86.ActiveCfg = Debug|Win32
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Debug|x86.Build.0 = Debug|Win32
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Release|x64.ActiveCfg = Release|x64
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Release|x64.Build.0 = Release|x64
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Release|x86.ActiveCfg = Release|Win32
+		{63B5F7A0-B90D-4BBC-BBCB-C61B25054911}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
